+++ src/main/java/com/bizcub/randomItemSpeedrun/Main.java
@@ -3,10 +3,10 @@ package com.bizcub.randomItemSpeedrun;
 import com.bizcub.randomItemSpeedrun.config.Compat;
 import com.bizcub.randomItemSpeedrun.config.Configs;
 import com.bizcub.randomItemSpeedrun.gui.Speedrun;
+import com.bizcub.randomItemSpeedrun.util.Constants;
+import com.bizcub.randomItemSpeedrun.util.RemovableItems;
 import com.bizcub.randomItemSpeedrun.util.Utils;
-import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.world.item.Item;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -20,20 +20,30 @@ public class Main {
         if (Compat.isClothConfigLoaded()) Configs.init();
         Utils.readSpeedruns();
         game = new Game();
-
-        fillItemsList();
-        removeImpossibleItems();
+        setDifficulty();
     }
 
-    private static void fillItemsList() {
-        Registry<Item> itemRegistry = BuiltInRegistries.ITEM;
-        var items = itemRegistry.stream().toList();
-        items.forEach(item -> allItemsId.add(Utils.convertComponentToId(item.getDescriptionId())));
+    public static void setDifficulty() {
+        if (Compat.isClothConfigLoaded()) {
+            switch (Configs.getInstance().difficulty) {
+                case EASY -> fillItemsList(Constants.notEasyItems());
+                case NORMAL -> fillItemsList(Constants.notMediumItems());
+                case HARD -> fillItemsList(Constants.notHardItems());
+                case HARDCORE -> fillItemsList(Constants.impossibleItems());
+                default -> {}
+            }
+        } else {
+            fillItemsList(Constants.notMediumItems());
+        }
     }
 
-    private static void removeImpossibleItems() {
-        List<String> removableItems = List.of("player_head", "air", "debug_stick", "bedrock", "barrier", "jigsaw", "light", "end_portal_frame", "reinforced_deepslate", "farmland", "vault", "suspicious_sand", "suspicious_gravel", "knowledge_book");
-        allItemsId.removeAll(removableItems);
-        allItemsId.removeIf(id -> id.contains("command_block") || id.contains("test") || id.contains("structure") || id.contains("infested") || id.contains("spawn_egg") || id.contains("spawner"));
+    private static void fillItemsList(RemovableItems items) {
+        allItemsId.clear();
+        BuiltInRegistries.ITEM.stream().toList().forEach(item ->
+                allItemsId.add(Utils.convertComponentToId(item.getDescriptionId())));
+
+        allItemsId.removeAll(items.equalItems());
+        for (String item : items.containItems())
+            allItemsId.removeIf(id -> id.contains(item));
     }
 }