+++ src/main/java/com/bizcub/randomItemSpeedrun/platform/Fabric.java
@@ -9,6 +9,7 @@ import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
+import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
 import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
 import net.minecraft.client.Minecraft;
 
@@ -20,9 +21,15 @@ public class Fabric implements ClientModInitializer {
 
         HudRenderCallback.EVENT.register((guiGraphics, deltaTracker) -> Utils.renderHud(guiGraphics));
 
+        KeyBindingHelper.registerKeyBinding(Constants.OPEN_SCREEN);
+        KeyBindingHelper.registerKeyBinding(Constants.QUICK_START);
+
         ClientTickEvents.END_CLIENT_TICK.register(client -> {
-            while (Constants.MY_KEYBIND.consumeClick()) {
+            while (Constants.OPEN_SCREEN.consumeClick()) {
                 Minecraft.getInstance().setScreen(new GameStartScreen());
+            }
+            while (Constants.QUICK_START.consumeClick()) {
+                Main.game.buttonPressed();
             }
         });
     }