+++ src/main/java/com/bizcub/randomItemSpeedrun/util/RemovableItems.java
@@ -1,0 +1,9 @@
+package com.bizcub.randomItemSpeedrun.util;
+
+import java.util.List;
+
+public record RemovableItems(
+        List<String> equalItems,
+        List<String> containItems
+) {
+}