+++ CHANGELOG.md
@@ -1,1 +1,8 @@
-* Block objects in versions 1.21.5+ no longer appear too pixelated
+* Added difficulty selection
+  * Easy - items that don't require entering the Nether
+  * Normal - items that don't require entering the End
+  * Hard - all items except those that take too long
+  * Hardcore - all items
+* Updated list of impossible items
+* Added a keybind for faster speedruns
+* On earlier versions, the mod settings icon now has a size dependent on the screen size